Add `gpctl api workspaces ownertoken` command
